==> 02_detail_download.ipynb <==
from webbot import Browser
import pandas as pd
import re
import time
import random

web = Browser()
web.go_to('https://www.strava.com/login') 
web.type('mkarkalik@gmail.com' , into='email')
web.type('akosamas' , into='password' , id='password') # specific selection
web.click('submit', id = 'login-button') # you are logged in ^_^

regex_watch = r'(?<=device spans8)">(.*)(?=</div><div class="gear spans8)'
regex_shoes = r'(?<=Shoes:<span class="gear-name">)(.*)(?=</span></div></div></div></div></div><footer class="achievements row">)'
regex_distance = r'(?<=inline-stats section"><li><strong>)(.*)(?=<abbr class="unit" title="kilometers">km)'

race_table = pd.read_csv('barcelona_half_2019_raw.csv', encoding = "ISO-8859-1")

watch_list = []
shoes_list = []
distance_list = []
i = 0

for url in race_table['race_url']:
    #time.sleep(random.randint(1,4))
    web.go_to('https://www.strava.com'+url)
    activity = web.get_page_source()
    activity_clear = activity.replace('\n','')
    watch = re.findall(regex_watch, activity_clear)
    shoes = re.findall(regex_shoes, activity_clear)
    distance = re.findall(regex_distance, activity_clear)
    watch_list.append(watch)
    shoes_list.append(shoes)
    distance_list.append(distance)
    i = i+1

pd.DataFrame(watch_list).to_csv('watch_list.csv')
pd.DataFrame(shoes_list).to_csv('shoes_list.csv')
pd.DataFrame(distance_list).to_csv('distance_list.csv')

#'Too Many Requests' create condition

# Different aproach

from requests import Session
from bs4 import BeautifulSoup as bs
 
with Session() as s:
    site = s.get("https://www.strava.com/login")
    bs_content = bs(site.content, "html.parser")
    token1 = bs_content.find("meta", {"name":"csrf-token"})["content"]
    token2 = bs_content.find("input", {"name":"authenticity_token"})["value"]
    login_data = {"email":"mkarkalik@gmail.com", 
                  "password":"akosamas", 
                  "csrf-token":token1,
                  "authenticity_token":token2}
    s.post("https://www.strava.com/login", login_data)
    home_page = s.get("https://www.strava.com/")
    print(bs(home_page.content, "html.parser"))
    #print(token2)

race_table['Watch'] = watch_list
race_table['Shoes'] = shoes_list
race_table['Distance'] = distance_list

race_table.to_csv('barcelona_half_2019.csv')
